git monitor commit.

+++ nemu/src/cpu/instr/ret.c
@@ -2,7 +2,7 @@
 
 make_instr_func(ret_near)
 {
-    uint32_t len=1;
+    uint8_t len=1;
     opr_dest.data_size=data_size;//!!initialize
     opr_dest.type=OPR_MEM;
     opr_dest.addr=cpu.esp;
